requirements.txt
alpaca-py
fastapi
uvicorn
pydantic
flake8
black
pre-commit
pytest
pytest-mock
pytest-asyncio 